(feature): add cmds to jump to top, bottom and center of view
Adds:
* high - changes selected file to the file at the top of the visible window.
* middle - like `high`, but goes to the file directly between `high` and `low`.
* low - like `high`, but goes to the last file in the visible window.

NOTE: middle, by design, rounds down when an odd number of files are
available. This means if you have 5 files, middle will take you to the
third.

NOTE: all of these methods work only on the visible region of files.
If you for example goto the last file in a directory, run
`recenter-top` to make it the only visible file and then try running
any of the above commands, it'll have no affect.