(icons): allow complex icons
This patch gives lf the ability to:
- Apply icons for folders the same way as it does for files
- Specify icons with none-extension based globs (Eg. `README=foo`, `*webpack*=bar`).

It also enforces a few new considerations:
1. LF_ICONS prioritises earlier matches over later ones. `*.txt=foo:*.txt=bar` will
   always return foo, never bar.
2. LF_ICONS prioritises globs over types. `tw=foo:*.txt=bar` will consider `*.txt`
   before `tw` even though it appears earlier in LF_ICONS.

Aside from that I've tried to maintain backwards compatibility with the existing
implementation, crucially in regards to performance. If you only ever use the
extension-globs and type-patterns than this implementation should be just as
performant as the existing implementation.

For every none-extension based glob you use an extra regex check will be performed,
but only on these patterns that exist before a matching extension. For example: If
LF_ICONS is:
- `*.txt=foo`
- `*t*=bar`

Any .txt icon lookups will never check the pattern `*t*` because `*.txt` takes
precedence (due to occurring earlier in LF_ICONS). Conversely if LF_ICONS is:
- `*f*=bar`
- `*b*=bar`
- `*txt*=baz`
- `*.txt=bag`

We'll be checking all 3 of the patterns `*f*`, `*b*`, `*txt*` until either a match is
found or we've reached the position of `*.txt` (in which case bag is used as the
icon).

Lastly if no glob lookup resolves to an icon, we compare file-types (see fileIconTypes)
or return " ".